# Clean up generation logic and output of StringMatcher