This PR aims to improve the situation mentioned [here](https://github.com/sebastienros/esprima-dotnet/pull/314#discussion_r948265709) by utilizing the string matcher source generator.

Also does a bit of cleanup around that.