@lahma 2022-08-18
Seems that this has the same outcome as my weekend experiment, things get slower. Oddly there's no memory benefit either.

## Esprima.Benchmark.FileParsingBenchmark

| **Diff**|Method|FileName|Mean|Error|Gen 0|Gen 1|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|
| Old |ParseProgram|angular-1.2.5|10.470 ms|0.0248 ms|343.7500|156.2500|5,744 KB|
| **New** |	|	| **11.052 ms (+6%)** | **0.0511 ms** | **343.7500 (0%)** | **156.2500 (0%)** | **5,744 KB (0%)** |
| Old |ParseProgram|backbone-1.1.0|1.393 ms|0.0070 ms|52.7344|21.4844|888 KB|
| **New** |	|	| **1.510 ms (+8%)** | **0.0067 ms** | **52.7344 (0%)** | **21.4844 (0%)** | **888 KB (0%)** |
| Old |ParseProgram|jquery-1.9.1|8.244 ms|0.0167 ms|296.8750|140.6250|4,991 KB|
| **New** |	|	| **8.791 ms (+7%)** | **0.0159 ms** | **296.8750 (0%)** | **140.6250 (0%)** | **4,991 KB (0%)** |
| Old |ParseProgram|jquery.mobile-1.4.2|13.221 ms|0.0358 ms|468.7500|234.3750|7,677 KB|
| **New** |	|	| **14.005 ms (+6%)** | **0.0451 ms** | **468.7500 (0%)** | **234.3750 (0%)** | **7,677 KB (0%)** |
| Old |ParseProgram|mootools-1.4.5|6.600 ms|0.0088 ms|242.1875|117.1875|4,028 KB|
| **New** |	|	| **7.130 ms (+8%)** | **0.0224 ms** | **242.1875 (0%)** | **117.1875 (0%)** | **4,028 KB (0%)** |
| Old |ParseProgram|underscore-1.5.2|1.173 ms|0.0063 ms|44.9219|19.5313|748 KB|
| **New** |	|	| **1.278 ms (+9%)** | **0.0016 ms** | **44.9219 (0%)** | **19.5313 (0%)** | **748 KB (0%)** |
| Old |ParseProgram|yui-3.12.0|5.974 ms|0.0084 ms|226.5625|85.9375|3,789 KB|
| **New** |	|	| **6.406 ms (+7%)** | **0.0222 ms** | **226.5625 (0%)** | **85.9375 (0%)** | **3,789 KB (0%)** |