@lahma 2022-08-18
For reference, this is the dictionary I used, it contains all JS keywords too which were missing, like `async`.

```csharp
    [StringMatcher(
        "if", "in", "do", "var", "for", "new", "try", "let", "this", "else", "case", "void", "with", "enum", "await",
        "while", "break", "catch", "throw", "const", "yield", "class", "super", "return", "typeof", "delete", "switch",
        "export", "import", "default", "finally", "extends", "function", "continue", "debugger", "instanceof", "async", "static",
        "undefined", "true", "false", "null", "get", "set",
        // some common ones in our test data set (benchmarks + test suite)
        "length", "object", "Object", "obj", "Array", "Math", "data", "done", "args", "arguments", "Symbol", "prototype",
        "options", "value", "name", "self", "&&", "||", "===", "!===", "key", "\"use strict\"", "use strict"
    )]
    private static partial string? Slice(ReadOnlySpan<char> span);
```