@lahma 2022-08-18
This is a competing solution for both your earlier pooling and my (unfinished StringSegment/TextSpan). One potential I know that is here, is the fact that we should be able to match keywords etc with *reference equality* as long as they are part of the dictionary. So instead of  `token.Type == Keyword && "let".Equals(token.Value)` we could write `token.Type == Keyword && ReferenceEquals("let", token.Value)` because the string should be interned by the compiler.

This solution will not reduce memory pressure, but *could* improve string equality checks which are pretty much the bread and butter. In my StringSegment PR it's easy to see that perf is getting slower with spanification, but memory usage takes jump downwards.