@lahma 2022-08-18
> This wasn't intended as a solution to the string allocation problem but merely as an attempt on improving the `ParserExtensions.Slice` function. According to your benchmark, unfortunately, it fails to do so: although code was simplified, execution time got worse. So I'll drop the first commit and let's merge only the other one which does some cleanup around the string matcher generator logic and output.

Yes, the code is a low cleaner, a good nerd-snipe for me to investigate which paths actually make slower than all the indexing - my guess would be large switches (above 6 targets) which translate to hash code.

> > One potential I know that is here, is the fact that we should be able to match keywords etc with reference equality as long as they are part of the dictionary. So instead of token.Type == Keyword && "let".Equals(token.Value) we could write token.Type == Keyword && ReferenceEquals("let", token.Value) because the string should be interned by the compiler.
> 
> Wouldn't this rather apply to the string pooling PR? I mean `token.Type == Keyword && "let".Equals(token.Value)` makes sense only when we've made sure that `token.Value` is an interned string instance (in the case of keywords)?

When `Slice` is used and dictionary contains all targets, we should get interned result and that is what is set into tokens I believe.

> > EDIT: there's also a path that acorn uses, add extra enum flag to token to directly say it's "let", "async" etc at parsing time when say, keyword is checked. This way during parsing it would be just number equality.
> 
> Not sure that it would worth the hassle if we used `ReferenceEquals` as you mentioned above.

Yes, again a competing solution. The enum flag would also work in world where we use spans instead of substrings / string instances.