@adams85 2022-08-18
> Yes, the code is a low cleaner, a good nerd-snipe for me to investigate which paths actually make slower than all the indexing - my guess would be large switches (above 6 targets) which translate to hash code.

I'm almost sure that large switches on span parameters cause the slowdown. There's some compiler magic which generates hash for these spans. Looks like character-based comparison beats that (at least, in our case).